+++ frontend/packages/core/src/Assets/index.tsx
@@ -1,0 +1,2 @@
+export * from "./emojis";
+export * from "./icons";
