+++ frontend/packages/core/src/DataDisplay/Table/stories/table.stories.tsx
