+++ frontend/packages/core/src/DataDisplay/index.tsx
@@ -1,0 +1,15 @@
+export * from "./chip";
+export {
+  AccordionRow,
+  MetadataTable,
+  Table,
+  TableCell,
+  TableRow,
+  TableRowAction,
+  TableRowActions,
+  TreeTable,
+} from "./Table";
+export { default as Code } from "./text";
+export { Typography } from "./typography";
+export { AvatarIcon, StatusIcon } from "./icon";
+export type { TypographyProps } from "./typography";
