+++ frontend/packages/core/src/Feedback/index.tsx
@@ -1,4 +1,5 @@
 export { Alert } from "./alert";
+export { Dialog, DialogActions, DialogContent } from "./dialog";
 export { default as Error } from "./error";
 export { default as Hint } from "./hint";
 export { Note, NotePanel } from "./note";
