+++ frontend/packages/core/src/Feedback/note.tsx
@@ -3,7 +3,7 @@ import { Paper } from "@mui/material";
 import type { AlertColor as Color } from "@mui/material/Alert";
 
 import { Grid } from "../Layout";
-import { Link } from "../link";
+import { Link } from "../Navigation";
 import { styled } from "../Utils";
 
 import { Alert } from "./alert";
