+++ frontend/packages/core/src/Feedback/stories/dialog.stories.tsx
@@ -2,10 +2,10 @@ import React, { useState } from "react";
 import { action } from "@storybook/addon-actions";
 import type { Meta } from "@storybook/react";
 
+import { Button } from "../../Input";
+import TextField from "../../Input/text-field";
 import type { DialogProps } from "../dialog";
 import { Dialog, DialogActions, DialogContent } from "../dialog";
-import { Button } from "../Input";
-import TextField from "../Input/text-field";
 
 export default {
   title: "Core/Dialog",
