+++ frontend/packages/core/src/NPS/feedback.tsx
@@ -5,13 +5,13 @@ import { debounce } from "lodash";
 import { v4 as uuid } from "uuid";
 
 import { userId } from "../AppLayout/user";
+import { Typography } from "../DataDisplay/typography";
 import { Alert } from "../Feedback";
 import type { SelectOption } from "../Input";
 import { Button, Select, TextField } from "../Input";
 import { Grid } from "../Layout";
 import { client } from "../Network";
 import type { ClutchError } from "../Network/errors";
-import { Typography } from "../typography";
 import { styled } from "../Utils";
 
 import EmojiRatings, { Rating } from "./emojiRatings";
