+++ frontend/packages/core/src/Navigation/stories/not-found.stories.tsx
