+++ frontend/packages/core/src/Navigation/stories/stepper.stories.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 import type { Meta } from "@storybook/react";
 
-import { Button, ButtonGroup } from "../Input";
+import { Button, ButtonGroup } from "../../Input";
+import { styled } from "../../Utils";
 import type { StepperProps } from "../stepper";
 import { Step, Stepper } from "../stepper";
-import { styled } from "../Utils";
 
 const Text = styled("div")({
   textAlign: "center",
