+++ frontend/packages/core/src/Navigation/stories/tabs.stories.tsx
