+++ frontend/packages/core/src/Navigation/tab.tsx
@@ -3,7 +3,7 @@ import { TabContext, TabList, TabPanel as MuiTabPanel } from "@mui/lab";
 import type { TabProps as MuiTabProps, TabsProps as MuiTabsProps } from "@mui/material";
 import { Tab as MuiTab } from "@mui/material";
 
-import { styled } from "./Utils";
+import { styled } from "../Utils";
 
 const StyledTab = styled(MuiTab)({
   minWidth: "111px",
