+++ frontend/packages/core/src/Surface/index.tsx
@@ -6,6 +6,6 @@ export {
   AccordionGroup,
 } from "./accordion";
 export { Card, CardContent, CardHeader } from "./card";
+export { default as ExpansionPanel } from "./panel";
 export { default as Paper } from "./paper";
-
 export type { CardHeaderSummaryProps } from "./card";
