+++ frontend/packages/core/src/Utils/flags.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import type { clutch as IClutch } from "@clutch-sh/api";
 
-import type { ClutchError } from "./Network/errors";
-import { client } from "./Network/index";
+import type { ClutchError } from "../Network/errors";
+import { client } from "../Network/index";
 
 const FEATURE_FLAG_POLL_RATE = +process.env.REACT_APP_FF_POLL || 300000;
 const FF_CACHE_TTL = +process.env.REACT_APP_FF_CACHE_TTL || 60000;
