+++ frontend/packages/core/src/Utils/index.tsx
@@ -1,5 +1,7 @@
+export { FeatureOff, FeatureOn, SimpleFeatureFlag } from "./flags";
 export { Popper, PopperItem } from "./popper";
 export { default as styled } from "./styled";
+export { default as TimeAgo } from "./timeago";
 export { css as EMOTION_CSS, keyframes as EMOTION_KEYFRAMES } from "@emotion/react";
 
 export type { StyledComponent } from "@emotion/styled";
