+++ frontend/packages/core/src/index.tsx
@@ -1,49 +1,21 @@
 export { userId } from "./AppLayout/user";
-export * from "./Assets/emojis";
-export * from "./Assets/icons";
+export { default as ClutchApp } from "./AppProvider";
+export * from "./Assets";
 export * from "./Charts";
-export * from "./chip";
 export { default as Confirmation } from "./confirmation";
 export { useWorkflowStorageContext, useWizardContext, WizardContext } from "./Contexts";
-export { Dialog, DialogActions, DialogContent } from "./dialog";
+export * from "./DataDisplay";
 export * from "./Feedback";
-export { FeatureOff, FeatureOn, SimpleFeatureFlag } from "./flags";
-export { AvatarIcon, StatusIcon } from "./icon";
 export * from "./Input";
-export { Link } from "./link";
+export * from "./Layout";
 export { default as Loadable } from "./loading";
-export {
-  convertSearchParam,
-  useLocation,
-  useNavigate,
-  useParams,
-  useSearchParams,
-} from "./navigation";
+export * from "./Navigation";
 export { client } from "./Network";
 export * from "./NPS";
-export { default as ExpansionPanel } from "./panel";
 export { default as Resolver } from "./Resolver";
-export { Step, Stepper } from "./stepper";
 export * from "./Surface";
 export * from "./Utils";
-export { Tab, Tabs } from "./tab";
-export {
-  AccordionRow,
-  MetadataTable,
-  Table,
-  TableCell,
-  TableRow,
-  TableRowAction,
-  TableRowActions,
-  TreeTable,
-} from "./Table";
-export { default as Code } from "./text";
-export { default as TimeAgo } from "./Utils/timeago";
-export { Typography } from "./typography";
-export { default as ClutchApp } from "./AppProvider";
-export * from "./Layout";
 
 export type { BaseWorkflowProps, WorkflowConfiguration } from "./AppProvider/workflow";
 export type { WorkflowRemoveDataFn, WorkflowRetrieveDataFn, WorkflowStoreDataFn } from "./Contexts";
 export type { ClutchError } from "./Network/errors";
-export type { TypographyProps } from "./typography";
